requirements.txt
Django==4.0
djangorestframework==3.13.1
Pillow==9.0.0
drf-yasg==1.20.0
psycopg2==2.9.3
python-dotenv==0.19.2